platform: Make platform interface more opaque
There shouldn't be any need for the platform structure to be visible to
the core. Unfortunately, the PALDevData{Get,Set} macros break that
(Dear ST: Why? Why not just add an API for "GetDevData"? Seems silly,
but not my code).

So, reduce VL53L0X_Dev to the bare minumum, and we'll subclass it in the
platform code. A bit more ugly, but better for separation.